% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_angle.R
\name{cal_angle}
\alias{cal_angle}
\alias{cal_azimuth}
\title{Rotation angle of two speed vector}
\usage{
cal_angle(v1, v2, clockwise = -1)

cal_azimuth(v, refer = c(0, 1))
}
\arguments{
\item{v1, v2}{speed vector, in deg, (lon, lat).
\itemize{
\item \code{v1} is the previous time.
\item \code{v2} is the current time;
}}

\item{clockwise}{one of \code{c(-1, 1)}:
-\code{clockwise = -1}: clockwise corresponds to negative angle
-\code{clockwise =  1}: clockwise corresponds to positive angle (azimuth angle in this style)}
}
\value{
\strong{v2 -> v1}:
\itemize{
\item rotation angle (in deg, \verb{[-180, 180]})
\itemize{
\item counterclockwise is positive
\item clockwise is negative
}
\item azimuth angle (in deg, \verb{[-180, 180]})
\itemize{
\item the angle with north (\code{c(0, 1)})
}
}
}
\description{
Rotation angle of two speed vector
}
\examples{
cal_angle(c(1, 0), c(2, 0)) # == 0
cal_angle(c(0, 1), c(0, 2)) # == 0

p1 <- c(1, 0)
p2 <- c(0, 1)

cal_angle(p1, p2) # == 90, counterclockwise
cal_angle(p2, p1) # == -90, clockwise
}
