% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_pressure.R
\name{cal_ea}
\alias{cal_ea}
\title{Estimate actual vapor pressure.}
\usage{
cal_ea(tmin, tmax = NULL, rhmax = NULL, rhmean = NULL, rhmin = NULL)
}
\arguments{
\item{tmin}{Daily minimum air temperature at 2m height \verb{[deg Celsius]}.}

\item{tmax}{Daily maximum air temperature at 2m height \verb{[deg Celsius]}.}

\item{rhmax, rhmean, rhmin}{Daily max, mean and min relative humidity \verb{[\%]}.}
}
\value{
Actual vapor pressure (i.e. avp or ea) \verb{[kPa]}.
}
\description{
Estimate actual vapor pressure by providing daily maximum and minimum
air temperature, daily maximum, mean and minimum relative humidity.
}
\details{
tmin must be provided.
If tmax was not provided, avp will only estimated by tmin. If both rhmax and
rhmin are provided, avp will be estimated by tmax, tmin, rhmax and rhmin.

If rhmin was not provided but provided rhmean, avp will estimated by rhmean,
tmax and tmin.

If rhmin and rhmean are not provided, avp will estimated by rhmax, tmax and
tmin.

If rhmax, rhmean and rhmin are not provided, avp will only estimated by tmin.
}
