% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ET0_penman65.R
\name{cal_gamma}
\alias{cal_gamma}
\title{Calculate psychrometric constant.}
\usage{
cal_gamma(pres = NULL, z = NULL)
}
\arguments{
\item{pres}{Air pressure \verb{[kPa]}.}

\item{z}{Elevation above sea level \verb{[m]}. Must provided if pres
are not provided.}
}
\value{
Psychrometric constant \verb{[kPa degC-1]}.
}
\description{
Calculate psychrometric constant.
}
