% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_radiation.R
\name{cal_Ra}
\alias{cal_Ra}
\alias{ext_rad}
\title{Daily extraterrestrial radiation.}
\usage{
cal_Ra(lat, J)

ext_rad(lat, J)
}
\arguments{
\item{lat}{Latitude \verb{[degree]}.}

\item{J}{\code{doy} vector.}
}
\value{
extraterrestrial radiation \verb{[MJ m-2 day-1]}.
}
\description{
Estimate daily extraterrestrial radiation \verb{[MJ m-2 day-1]}.
}
\seealso{
\code{\link[=cal_Rs]{cal_Rs()}}
}
