% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_radiation_Ta.R
\name{radiation_Ta}
\alias{radiation_Ta}
\alias{cal_radiation_Ta}
\title{air temperature under solar radiation}
\usage{
radiation_Ta(Rs, Rl = 0, albedo = 0.3, emiss = 0.97)

cal_radiation_Ta(
  Ta,
  Rs,
  Rl = 0,
  dt = 3600,
  method = c("approx", "exact"),
  ...,
  albedo = 0.3,
  emiss = 0.97
)
}
\arguments{
\item{Rs, Rl}{numeric vector, inward shortwave and longwave radiation, in \code{W/m2}}

\item{Ta}{scalar, the initial air temperature, in \code{degC}}

\item{dt}{delta \code{t}, in second}
}
\description{
air temperature under solar radiation
}
