% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_radiation_Ta.R
\name{cal_Rl_out}
\alias{cal_Rl_out}
\title{black body radiation}
\usage{
cal_Rl_out(Ta, emiss = 0.97)
}
\arguments{
\item{Ta}{temperature in C deg}

\item{emiss}{Emissivity coefficient}
}
\value{
radiation in W m-2
}
\description{
black body radiation
}
