% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_radiation.R
\name{cal_Rlin}
\alias{cal_Rlin}
\title{Estimate Incoming longwave radiation.}
\usage{
cal_Rlin(temp, ea = NULL, s = 1, method = "KON")
}
\arguments{
\item{temp}{Near surface air temperature \verb{[deg Celsius]}.}

\item{ea}{Near surface actual vapour pressure \verb{[kPa]}.}

\item{s}{Cloud air emissivity, the ratio between acutal incomming shortwave
radiation and clear sky incoming shortwave radiation.}

\item{method}{The method to esteimate the air emissivity. Must be one of
'MAR', 'SWI', 'IJ', 'BRU', 'T', and 'KON'.}
}
\value{
Incomming longwave radiation \verb{[W /m2]}.
}
\description{
Estimate Incoming longwave radiation. Not included in FAO56
paper but added for convinience.
}
