% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_radiation.R
\name{cal_Rln}
\alias{cal_Rln}
\title{Estimate net outgoing longwave radiation.}
\usage{
cal_Rln(tmax, tmin, ea, Rs = NULL, Rso = NULL, cld = NULL)
}
\arguments{
\item{tmax}{Daily maximum air temperature at 2m height \verb{[deg Celsius]}.}

\item{tmin}{Daily minimum air temperature at 2m height \verb{[deg Celsius]}.}

\item{ea}{Actual vapor pressure \verb{[kPa]}. Can be estimated by maximum or minimum
air temperature and mean relative humidity.}

\item{Rs}{Incoming shortwave radiation at crop surface \verb{[MJ m-2 day-1]}.}

\item{Rso}{Clear sky incoming shortwave radiation, i. e. extraterrestrial
radiation multiply by clear sky transmissivity (i. e. a + b,
a and b are coefficients of Angstrom formula. Normally 0.75)
\verb{[MJ m-2 day-1]}.}

\item{cld}{Cloud cover \verb{[fraction]}. If provided it would be directly used to
calculate net outgoing longwave radiation than Rso.}
}
\value{
Net outgoing longwave radiation \verb{[MJ m-2 day-1]}
}
\description{
Estimate net outgoing longwave radiation.
}
