% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_radiation.R
\name{cal_Rs}
\alias{cal_Rs}
\title{Daily inward shortwave solar radiation.}
\usage{
cal_Rs(lat, J, ssd = NULL, cld = NULL, Z = 0, albedo = 0.23, a = 0.25, b = 0.5)
}
\arguments{
\item{lat}{Latitude \verb{[degree]}.}

\item{ssd}{sunshine duration \link{hours}. If \code{ssd = NULL}, \code{Rs} is the clear-sky
solar radiation.}

\item{cld}{Cloud cover \verb{[fraction]}. If provided it would be directly used to
calculate solar radiation rather than SSD and parameter a and b.}

\item{a}{Coefficient of the Angstrom formula. Determine the relationship
between ssd and radiation. Default 0.25.}

\item{b}{Coefficient of the Angstrom formula. Default 0.50.}

\item{dates}{A R Date type of a vector of Date type. If not provided, it will
regard the ssd series is begin on the first day of a year.}
}
\value{
A data.table, solar radiation at crop surface \verb{[MJ m-2 day-1]}.
\itemize{
\item \code{Ra}: Clear-sky surface downward shortwave radiation
\item \code{Rs}: Surface downward shortwave radiation
}
}
\description{
Daily inward shortwave solar radiation at crop surface \verb{[MJ m-2 day-1]} by
providing sunshine duration (SSD) in hours or cloud cover in fraction.
}
\references{
\enumerate{
\item Martinezlozano J A, Tena F, Onrubia J E, et al. The historical
evolution of the Angstrom formula and its modifications: review and
bibliography.\link{J}. Agricultural & Forest Meteorology, 1985, 33(2):109-128.
\item https://github.com/sbegueria/SPEI/blob/master/R/penman.R
}
}
\seealso{
\code{\link[=cal_Ra]{cal_Ra()}}
}
