% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_radiation.R
\name{cal_Rsi_toa}
\alias{cal_Rsi_toa}
\title{Daily extraterrestrial radiation (top of atmosphere)}
\usage{
cal_Rsi_toa(lat, J)
}
\arguments{
\item{lat}{Latitude \verb{[degree]}.}

\item{J}{\code{doy} vector.}
}
\value{
extraterrestrial radiation \verb{[MJ m-2 day-1]}.
}
\description{
Estimate daily extraterrestrial radiation \verb{[MJ m-2 day-1]}.
}
\examples{
cal_Rsi_toa(20, 1)
# [1] 25.84874
}
\seealso{
\code{\link[=cal_Rsi]{cal_Rsi()}}
}
