% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_sun_angle.R
\name{cal_ssd}
\alias{cal_ssd}
\alias{cal_sunset_angle}
\alias{ws2ssd}
\title{sunset hour angle}
\usage{
cal_sunset_angle(lat, J)

ws2ssd(ws)

cal_ssd(lat, J)
}
\arguments{
\item{lat}{Latitude \verb{[degree]}.}

\item{J}{\code{doy} vector.}

\item{ws}{sunset hour angle}
}
\description{
Calculating sunset hour angle (\code{ws}) according to Allen Eq. 25.
}
