% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_Ts.R
\name{cal_Tw}
\alias{cal_Tw}
\title{wetbulb temperature}
\usage{
cal_Tw(ea, Tair, Pa = atm)
}
\arguments{
\item{ea}{Actual vapor pressure (kPa)}

\item{Tair}{2m air temperature (degC)}

\item{Pa}{surface air pressure (kPa)}
}
\description{
wetbulb temperature
}
