% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_Ts.R
\name{cal_wetbulb}
\alias{cal_wetbulb}
\alias{cal_Tw}
\title{wetbulb temperature}
\usage{
cal_wetbulb(Tair, Tdew, Pa = atm)

cal_Tw(ea, Tair, Pa = atm)
}
\arguments{
\item{Tair}{2m air temperature (degC)}

\item{Pa}{surface air pressure (kPa)}

\item{ea}{Actual vapor pressure (kPa)}
}
\description{
wetbulb temperature
}
\references{
\enumerate{
\item https://github.com/wrf-model/WRF/blob/master/phys/module_diag_functions.F#L1154
}
}
