% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temperature_unit_convertion.R
\name{convert_temperature}
\alias{convert_temperature}
\alias{celsius.to.fahrenheit}
\alias{celsius.to.kelvin}
\alias{fahrenheit.to.celsius}
\alias{fahrenheit.to.kelvin}
\alias{kelvin.to.celsius}
\alias{kelvin.to.fahrenheit}
\title{Convert from one temperature metric to another}
\usage{
convert_temperature(temperature, old_metric, new_metric)

celsius.to.fahrenheit(T.celsius)

celsius.to.kelvin(T.celsius)

fahrenheit.to.celsius(T.fahrenheit)

fahrenheit.to.kelvin(T.fahrenheit)

kelvin.to.celsius(T.kelvin)

kelvin.to.fahrenheit(T.kelvin)
}
\arguments{
\item{temperature}{A numeric vector of temperatures to be converted.}

\item{old_metric}{The metric from which you want to convert. Possible options are:
\itemize{
\item \code{fahrenheit}, \code{f}
\item \code{kelvin}, \code{k}
\item \code{celsius}, \code{c}
}}

\item{new_metric}{The metric to which you want to convert. The same options
are possible as for \code{old_metric}.}

\item{T.kelvin, T.fahrenheit, T.celsius}{Numeric vector of temperatures in Kelvin, Fahrenheit and Celsius.}
}
\value{
A numeric vector with temperature converted to the metric specified
by the argument \code{new_metric}.

A numeric vector of temperature values in the new unit.
}
\description{
This function allows you to convert a vector of temperature values between
Fahrenheit, Celsius, and degrees Kelvin.
}
\examples{
data(lyon)
lyon$TemperatureF <- convert_temperature(lyon$TemperatureC,
    old_metric = "c", new_metric = "f"
)
lyon

data(norfolk)
norfolk$TempC <- convert_temperature(norfolk$TemperatureF,
    old_metric = "f", new_metric = "c"
)
norfolk

data(angeles)
angeles$TemperatureC <- convert_temperature(angeles$TemperatureK,
    old_metric = "kelvin", new_metric = "celsius"
)
angeles
# Convert from Celsius
data(lyon)
lyon$TemperatureF <- celsius.to.fahrenheit(lyon$TemperatureC)
lyon$TemperatureK <- celsius.to.kelvin(lyon$TemperatureC)
lyon
# Convert from Fahrenheit
data(norfolk)
norfolk$TempC <- fahrenheit.to.celsius(norfolk$TemperatureF)
norfolk$TemperatureK <- fahrenheit.to.kelvin(norfolk$TemperatureF)
norfolk
# Convert from Kelvin
data(angeles)
angeles$TemperatureC <- kelvin.to.celsius(angeles$TemperatureK)
angeles$TemperatureF <- kelvin.to.fahrenheit(angeles$TemperatureK)
angeles
}
\references{
\enumerate{
\item \href{http://www.wpc.ncep.noaa.gov/html/heatindex.shtml}{online heat index calculator}
\item \href{http://www.srh.noaa.gov/epz/?n=wxcalc_tempconvert}{online temperature converter}
}
}
\author{
Joshua Ferreri \email{joshua.m.ferreri@gmail.com},
Brooke Anderson \email{brooke.anderson@colostate.edu},
Roger Peng \email{rdpeng@gmail.com}
}
