% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_tree.R
\name{correct_basin_net_shapefile}
\alias{correct_basin_net_shapefile}
\title{correct_basin_net_shapefile}
\usage{
correct_basin_net_shapefile(
  ID,
  shp,
  tree,
  outdir = "basins/subbasin",
  prefix = "shed_珠江",
  overwrite = FALSE
)
}
\arguments{
\item{ID}{the ID of basins}

\item{shp}{basin shapefile, sf polygon object}

\item{tree}{returned by \code{\link[=plot_StreamNet]{plot_StreamNet()}}}

\item{outdir}{outdir of shapefile}

\item{prefix}{prefix of shapefile}
}
\description{
correct_basin_net_shapefile
}
