% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basin_flowdir.R
\name{create_flow_direction}
\alias{create_flow_direction}
\title{Create the flow direction of each gridcell for VIC routing model.}
\usage{
create_flow_direction(
  flow_file,
  domain_file,
  arc_code = TRUE,
  diag_tsh = 0.382
)
}
\arguments{
\item{flow_file}{An netCDF4 file of the flow accumulation data from high resolution DEM
(about 90m). Must containing all the area.}

\item{domain_file}{Domain file (netCDF4 format) of VIC model.}

\item{arc_code}{If the direction code is ArcInfo type. If FALSE, it would set to 1 to 8.}

\item{diag_tsh}{Diag threshould. Determine if flow to the diag gridcell.}
}
\value{
A list including the meta parameters (num of rows and columns, size of the cells,
x and y corner) and the matrix of the flow direction data.
}
\description{
Create the flow direction data of each VIC gridcell for the routing model
of VIC by providing a high resolution flow accumulation data (about 90m
usually caculated from SRTM DEM by ArcGIS) and the domain file of VIC model.
WARNING: This is a simple heuristic method and more fit for small gridcell.
It's usually needs further manual calibration by set_direc and plot_flow_direc.
}
