% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_pressure.R
\name{delta_es}
\alias{delta_es}
\title{The slope of the saturation vapour pressure curve.}
\usage{
delta_es(t)
}
\arguments{
\item{t}{Daily mean air temperature at 2m height \verb{[deg Celsius]}.}
}
\value{
Slope of the saturation vapour pressure curve.
}
\description{
the slope of the saturation vapour pressure curve by providing a air temperature.
}
