% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basin.R
\name{detect_basin}
\alias{detect_basin}
\title{Find out the gridcells in the basin controled by the station.}
\usage{
detect_basin(
  dir,
  stn_col,
  stn_row,
  arc_code = TRUE,
  xcor = NULL,
  ycor = NULL,
  csize = NULL,
  get_coords = FALSE
)
}
\arguments{
\item{dir}{Direction grid created by create_flow_direction() or a matrix.}

\item{stn_col}{Column of the hydrological station}

\item{stn_row}{Row of the hydrological station}

\item{arc_code}{If use the ArcInfo direction code. Default True.}

\item{xcor}{X coordinate of the corner of the grid. Should offered if dir is a matrix.}

\item{ycor}{Y coordinate of the corner of the grid. Should offered if dir is a matrix.}

\item{csize}{Size of the gridcells. Should offered if dir is a matrix.}

\item{get_coords}{If return the coordinates of the gridcells of the basin, or return the rows and columns.}
}
\value{
A data frame contains the coordinates of the gridcells in the basin.
}
\description{
Offer the grid of flow direction and the site (row and column)
of the hydrological station and returns a table of the coordinates
of the gridcells in the basin controled by the station.
}
