% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ET0_helper.R
\name{ET0_helper}
\alias{ET0_helper}
\alias{cal_U2}
\alias{cal_VPD}
\alias{cal_lambda}
\alias{cal_slope}
\alias{delta_es}
\alias{cal_gamma}
\alias{cal_bowen}
\alias{cal_Pa}
\title{helper functions for potential evapotranspiration (ET0)}
\usage{
cal_U2(Uz, z.wind = 10)

cal_VPD(Tair, Tdew)

cal_lambda(t)

cal_slope(t)

delta_es(t)

cal_gamma(Pa, lambda = 2.5)

cal_bowen(gamma, slope)

cal_Pa(z = NULL)
}
\arguments{
\item{z.wind}{Height where measure the wind speed \verb{[m]}. Default 10m.}

\item{t}{Air temperature \verb{[Celsius]}.}

\item{Pa}{land surface Air pressure \verb{[kPa]}.}

\item{z}{Elevation above sea level \verb{[m]}. Must provided if pres
are not provided.}
}
\description{
\itemize{
\item \code{lambda}: latent heat of vaporization, about \verb{[2.5 MJ kg-1]}.
\item \code{slope}: The slope of the saturation vapour pressure curve at certain air
temperature \code{t} (\link{deg Celsius}), \verb{[kPa degC-1]}.
\item \code{gamma}: psychrometric constant (\verb{[kPa degC-1]}), \code{Cp * Pa / (epsilon * lambda)}.
\item \code{U2}: 10m wind speed (m/s). According to wind profile relationship, convert U_z to U_2.
}
}
\references{
\enumerate{
\item Allen, R. G., & Luis S. Pereira. (1998). Crop
evapotranspiration-Guidelines for computing crop water requirements-FAO
Irrigation and drainage paper 56. European Journal of Agronomy, 34(3),
144–152. doi:10.1016/j.eja.2010.12.001
}
}
