% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ET0_models.R
\name{ET0_models}
\alias{ET0_models}
\alias{ET0_eq}
\alias{ET0_PM93}
\alias{ET0_PM98}
\alias{ET0_PT72}
\title{Potential Evapotranspiration models}
\usage{
ET0_eq(Rn, Tair, D, Pa = atm, ...)

ET0_PM93(Rn, Tair, D, Pa = atm, wind, z.wind = 10)

ET0_PM98(Rn, Tair, D, Pa = atm, wind, z.wind = 10, tall_crop = FALSE)

ET0_PT72(Rn, Tair, D, Pa = atm, alpha = 1.26, ...)
}
\arguments{
\item{Rn}{land surface net radiation, W m-2}

\item{Tair}{2m air temperature (degC)}

\item{D}{vapor pressure deficit (kPa)}

\item{Pa}{surface air pressure (kPa)}

\item{wind}{wind speed at the height of \code{z.wind}}

\item{z.wind}{Height where measure the wind speed \verb{[m]}. Default 10m.}
}
\description{
\itemize{
\item \code{ET0_PM93}: Penman equation simplified by Shuttleworth 1993
\item \code{ET0_PM98}: Penman-Monteith reference crop evapotranspiration, FAO56
}
}
\references{
\enumerate{
\item https://en.wikipedia.org/w/index.php?title=Penman_equation
\item Allen, R. G., & Luis S. Pereira. (1998). Crop
evapotranspiration-Guidelines for computing crop water requirements-FAO
Irrigation and drainage paper 56. European Journal of Agronomy, 34(3),
144–152. doi:10.1016/j.eja.2010.12.001
}
}
