% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basin.R
\name{fast_basin}
\alias{fast_basin}
\title{delineate hydrological basin according to AecGIS tools}
\usage{
fast_basin(dem = "dem.tif")
}
\arguments{
\item{dem}{the path of dem file}
}
\description{
\itemize{
\item add "c:/Python27/ArcGIS10.2/python2.exe" into system path
}
}
