% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_sun_angle.R
\name{get_localtime}
\alias{get_localtime}
\title{Get local time}
\usage{
get_localtime(time, lon, timeZone = 8)
}
\arguments{
\item{time}{POSIXct object}

\item{lon}{longitude in deg}
}
\description{
Get local time
}
