% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp.R
\name{grid2points}
\alias{grid2points}
\title{Convert the grid data (save as matrix) to points.}
\usage{
grid2points(
  grid,
  x = NULL,
  y = NULL,
  csize = NULL,
  xcor = NULL,
  ycor = NULL,
  NA_value = NULL,
  y_rev = FALSE
)
}
\arguments{
\item{grid}{A matrix of the gridded data.}

\item{x}{X coordinates for each column of the grid.}

\item{y}{Y coordinates for each column of the grid.}

\item{csize}{Size of each gridcell.}

\item{xcor}{Y coordinate of the southwest corner of the grid.}
}
\value{
A table of the points, including x and y coordinate and the values, converted
from the grid.
}
\description{
Conver the grid data to points, showed as a table, with the columns of
coordinates for each point and a column for their value.
}
