% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_runoff.R
\name{grid_date}
\alias{grid_date}
\title{grid_date}
\usage{
grid_date(x, col = "grey60", lty = 3, lwd = 0.4, ...)
}
\arguments{
\item{...}{others to \code{\link[=abline]{abline()}}}
}
\description{
grid_date
}
\keyword{internal}
