% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heat_index.R
\name{heat_index}
\alias{heat_index}
\alias{heat_index_julia}
\alias{.heat.index}
\alias{heat_index_vec}
\title{Algorithm for heat.index function.}
\usage{
heat_index(t, rh)

heat_index_julia(t, rh)

.heat.index(t = NA, rh = NA)

heat_index_vec(t = NA, rh = NA)
}
\arguments{
\item{t}{Numeric scalar of air temperature, in celsius.}

\item{rh}{Numeric scalar of relative humidity, in \%.}
}
\value{
A numeric scalar of heat index, in celsius
}
\description{
\code{heat_index} converts a numeric scalar of temperature
(in Fahrenheit) and a numeric scalar of relative humidity (in \%)
to heat index (in Fahrenheit). This function is not meant to be used
outside of the \code{\link[=heat.index]{heat.index()}} function.
}
\details{
If an impossible value of relative humidity is given
(below 0\% or above 100\%), heat index is returned as \code{NA}.
}
\note{
Equations are from the source code for the US National Weather
Service's
\href{http://www.wpc.ncep.noaa.gov/html/heatindex.shtml}{online heat index calculator}.
}
\examples{
n <- 1e3
t <- seq(-50, 100, length.out = n)
rh <- seq(0, 100, length.out = length(t))

## bug exist
i <- 88
heat_index(t[i], rh[i])
heat_index_vec(t[i], rh[i])
# heat_index_julia(t[i], rh[i])

julia_setup()
r <- heat_index(t, rh)
r_vec <- heat_index_vec(t, rh)
# r_jl <- heat_index_julia(t, rh)

# add tests for julia NA values
# r1 <- heat_index_julia(t*NA, rh*NA)

all.equal(r, r_vec)
# all.equal(r_jl, r_vec)

microbenchmark::microbenchmark(
    r2 <- heat_index(t, rh),
    r1 <- heat_index_julia(t, rh)
    # r_vec <- heat_index_vec(t, rh)
)
}
\references{
\enumerate{
\item Anderson GB, Bell ML, Peng RD. 2013. Methods to calculate the heat index
as an exposure Metric in environmental health research.
Environmental Health Perspectives 121(10):1111-1119.
\item National Weather Service Hydrometeorological Prediction
Center Web Team. Heat Index Calculator. 30 Jan 2015.
\url{http://www.wpc.ncep.noaa.gov/html/heatindex.shtml}.
Accessed 18 Dec 2015.
\item Rothfusz L. 1990. The heat index (or, more than you ever wanted to know
about heat index) (Technical Attachment SR 90-23). Fort Worth:
Scientific Services Division, National Weather Service.
\item R. Steadman, 1979. The assessment of sultriness. Part I: A
temperature-humidity index based on human physiology and clothing
science. Journal of Applied Meteorology, 18(7):861--873.
}
}
\author{
Brooke Anderson \email{brooke.anderson@colostate.edu},
Roger Peng \email{rdpeng@gmail.com}
}
