% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydroTools-package.R
\docType{package}
\name{hydroTools-package}
\alias{hydroTools}
\alias{hydroTools-package}
\title{hydroTools: Tools for hydrological model}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Dongdong Kong \email{kongdd.sysu@gmail.com} (\href{https://orcid.org/0000-0003-1836-8172}{ORCID})

Authors:
\itemize{
  \item Ruida Zhong \email{zrd2017@163.com} [contributor]
}

}
\keyword{internal}
