% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constant.R
\docType{data}
\name{Mw}
\alias{Mw}
\alias{Md}
\alias{epsilon}
\title{Molecular weight}
\format{
An object of class \code{numeric} of length 1.

An object of class \code{numeric} of length 1.

An object of class \code{numeric} of length 1.
}
\usage{
Mw

Md

epsilon
}
\description{
\code{Mw}: Molecular weight of dry air \eqn{M_d = 28.9634g/mol}
\code{Md}: \href{https://en.wikipedia.org/wiki/Molar_mass}{Molecular weight} of water vapor \eqn{M_w = 18.01528g/mol}
\code{epsilon}: the ratio of \code{Mw} to \code{Md}
}
\keyword{datasets}
