% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calib_XAJ.R
\name{plot_calib}
\alias{plot_calib}
\title{plot_calib}
\usage{
plot_calib(data, file = "output.pdf")
}
\arguments{
\item{data}{with the columns of \code{date}, \code{prcp}, \code{Q},  \code{Qobs}}
}
\description{
plot_calib
}
