% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basin_flowdir.R
\name{plot_flow_direc}
\alias{plot_flow_direc}
\title{Plot the flow direction of the flow direction data.}
\usage{
plot_flow_direc(
  direc,
  xlim = NULL,
  ylim = NULL,
  arc_code = TRUE,
  river_shp = NULL,
  row_rev = FALSE
)
}
\arguments{
\item{direc}{Flow direction data created by create_flow_direction().}

\item{arc_code}{If use the ArcInfo direction code.}

\item{river_shp}{Path of the river shp file.}
}
\description{
Plot the flow direction of a flow direction data (created by
create_flow_direction()) for visullization and calibration. Can
provide a shp file of the river network to plot helping the
mannual calibration.
}
