% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_tree.R
\name{plot_StreamNet}
\alias{plot_StreamNet}
\title{plot_tree}
\usage{
plot_StreamNet(info, title = "", fout = "tree.pdf", show = FALSE, root = -1)
}
\arguments{
\item{info}{A data.table with the column of \verb{["id", "iddown", "name"]}}

\item{title}{The title of the tree}
}
\description{
plot_tree
}
