% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp.R
\name{points2grid}
\alias{points2grid}
\title{Convert the point data (a table including columns of the coordinates and value) to
grid.}
\usage{
points2grid(points, x = NULL, y = NULL, val = NULL)
}
\arguments{
\item{x}{Vector of x coordinates or hich column store the x cordinate.}

\item{y}{Which column store the y cordinate.}

\item{val}{Which column store the point value.}

\item{grid}{A matrix of the gridded data.}
}
\value{
A matrix of the gridcell value.
}
\description{
Conver the point data, usually a table with columns including coordinates
and values to grids.
}
