% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constant.R
\docType{data}
\name{R}
\alias{R}
\alias{Rw}
\alias{Rd}
\alias{Cp}
\title{Specific gas constants}
\format{
An object of class \code{numeric} of length 1.

An object of class \code{numeric} of length 1.

An object of class \code{numeric} of length 1.

An object of class \code{numeric} of length 1.
}
\usage{
R

Rw

Rd

Cp
}
\description{
\itemize{
\item \code{R}: \href{https://en.wikipedia.org/wiki/Gas_constant}{gas constant}, J/(mol K)
\item \code{Rw}: \href{https://en.wikipedia.org/wiki/Gas_constant#Specific_gas_constant}{Specific gas constant}
of water vapor \eqn{R_w = \frac{1000R}{M_w} = 461.52J/(kg K)}.
\item \code{Rd}: Specific gas constant of dry air (J/(kg K)).
}
}
\keyword{datasets}
