% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_radiation.R
\name{rad_ext}
\alias{rad_ext}
\alias{cal_Ra}
\title{Daily extraterrestrial radiation.}
\usage{
rad_ext(lat, J)

cal_Ra(lat, J)
}
\arguments{
\item{lat}{Latitude \verb{[degree]}.}

\item{J}{\code{doy} vector.}
}
\value{
extraterrestrial radiation \verb{[MJ m-2 day-1]}.
}
\description{
Estimate daily extraterrestrial radiation \verb{[MJ m-2 day-1]}.
}
\examples{
cal_Ra(20, 1)
# [1] 25.84874
}
\seealso{
\code{\link[=cal_Rs]{cal_Rs()}}
}
