% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_radiation.R
\name{rad_surf}
\alias{rad_surf}
\alias{cal_Rs}
\title{Daily inward shortwave solar radiation.}
\usage{
rad_surf(lat, J, ssd = NULL, a = 0.25, b = 0.5, cld = NULL)

cal_Rs(lat, J, ssd = NULL, a = 0.25, b = 0.5, cld = NULL)
}
\arguments{
\item{lat}{Latitude \verb{[degree]}.}

\item{ssd}{sunshine duration \link{hours}. If \code{ssd = NULL}, \code{Rs} is the clear-sky
solar radiation.}

\item{a}{Coefficient of the Angstrom formula. Determine the relationship
between ssd and radiation. Default 0.25.}

\item{b}{Coefficient of the Angstrom formula. Default 0.50.}

\item{cld}{Cloud cover \verb{[fraction]}. If provided it would be directly used to
calculate solar radiation rather than SSD and parameter a and b.}

\item{dates}{A R Date type of a vector of Date type. If not provided, it will
regard the ssd series is begin on the first day of a year.}
}
\value{
Solar radiation at crop surface \verb{[MJ m-2 day-1]}.
}
\description{
Daily inward shortwave solar radiation at crop surface \verb{[MJ m-2 day-1]} by
providing sunshine duration (SSD) in hours or cloud cover in fraction.
}
\references{
Martinezlozano J A, Tena F, Onrubia J E, et al. The historical
evolution of the Angstrom formula and its modifications: review and
bibliography.\link{J}. Agricultural & Forest Meteorology, 1985, 33(2):109-128.
}
\seealso{
\code{\link[=cal_Ra]{cal_Ra()}}
}
