% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basin.R
\name{read_basin}
\alias{read_basin}
\title{Read the rows and columns of grids of a basin from a rout data file.}
\usage{
read_basin(data_path)
}
\arguments{
\item{data_path}{File path of the rou data file.}
}
\value{
A data frame contains the rows and columns of the gridcells in the basin.
}
\description{
Read the rows and columns of gridcells of a basin from a rout data file
created by VIC Hime.
}
