% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routing_muskingum.R
\name{routing_muskingum}
\alias{routing_muskingum}
\alias{routing_muskingum_nL}
\alias{calib_routing_muskingum_nL}
\title{routing_muskingum}
\usage{
routing_muskingum(I1, I2, Q1, deltaT, par)

routing_muskingum_nL(I, Q0, deltaT, par)

calib_routing_muskingum_nL(I, Q0, yobs, maxn = 1000)
}
\arguments{
\item{deltaT}{in hours}

\item{par}{\itemize{
\item \code{X}  : 0-1/2
\item \code{K}  : in the same unit as \code{deltaT}
\item \code{nt} : divide river into \code{nt} chunks (default 1)
}}
}
\description{
routing_muskingum
}
\author{
Dongdong Kong
}
