% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sceua.R
\name{sceua}
\alias{sceua}
\title{Shuffle Complex Evolution - University of Arizona (SCE-UA)}
\usage{
sceua(
  OFUN,
  pars,
  lower,
  upper,
  maxn = 10000,
  kstop = 5,
  pcento = 0.01,
  ngs = 5,
  npg = 5,
  nps = 5,
  nspl = 5,
  mings = 5,
  iniflg = 1,
  iprint = 0,
  iround = 3,
  peps = 1e-04,
  plog = rep(FALSE, length(pars)),
  implicit = NULL,
  timeout = NULL,
  ...
)
}
\arguments{
\item{OFUN}{objective function}

\item{pars}{starting values}

\item{lower}{lower bounds}

\item{upper}{upper bounds}

\item{maxn}{maximum number of iterations}

\item{kstop}{number of shuffling loops in which the criterion value must change
by the given percentage before optimization is terminated}

\item{pcento}{percentage by which the criterion value must change in given number of shuffling loops}

\item{ngs}{number of complexes in the initial population}

\item{npg}{number of points in each complex}

\item{nps}{number of points in a sub-complex}

\item{nspl}{number of evolution steps allowed for each complex before
complex shuffling}

\item{mings}{minimum number of complexes required, if the number of
complexes is allowed to reduce as the optimization proceeds}

\item{iniflg}{flag on whether to include the initial point in population
= 0, not included
= 1, included}

\item{iprint}{flag for controlling print-out after each shuffling loop
= 0, print information on the best point of the population
= 1, print information on every point of the population}

\item{plog}{whether optimization should be done in log10-domain. Either a single
TRUE value for all parameters, or a vector with TRUE/FALSE for the different parameters}

\item{implicit}{function for implicit boundaries (e.g. \code{sum(par[4]+par[5]) < 1})}

\item{npt}{total number of points in initial population (npt=ngs*npg)}
}
\description{
Calibration function which searches a parameter set which is minimizing the value of an objective function
}
\examples{
set.seed(1)
# generate example data from a function with three parameters
# with some random noise
fun = function(x, pars) pars[2]*sin(x*pars[1])+pars[3]
x = rnorm(50, sd = 3)
y = fun(x, pars = c(5, 2, 3)) +  rnorm(length(x), sd = 0.3)
plot(x,y)
   
# Objective function, summing up squared differences
OFUN = function(pars, x, yobs) {
  yvals = fun(x, pars)  
  sum((yvals-yobs)^2)
}

sceuares = sceua(OFUN, pars = c(0.1,0.1,0.1), lower = c(-10,0,-10), 
                 upper = c(10,10,10), x = x, yobs = y)
sceuares
xx = seq(min(x), max(x), 0.1)
lines(xx, fun(xx, pars = sceuares$par))
}
\references{
\enumerate{
\item Duan, Q., Sorooshian, S., and Gupta, V.K., 1992. Effective and efficient global
optimization for conceptual rainfall-runoff models. Water Resour. Res. 28 (4), 1015?1031.
\item Holland, H.H., 1975. Adaptation in natural and artificial systems,
University of Michigan Press, Ann Arbor.
\item Nelder, J.A. and Mead, R., 1965. A simplex method for function minimization,
Comput. J., 7(4), 308-313.
\item Price, W.L., 1987. Global optimization algorithms for a CAD workstation,
J. Optim. Theory Appl., 55(1), 133-146.
\item Skoien, J. O., Bloschl, G., Laaha, G., Pebesma, E., Parajka, J., Viglione, A., 2014.
Rtop: An R package for interpolation of data with a variable spatial support,
with an example from river networks. Computers & Geosciences, 67.
}
}
