% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basin_flowdir.R
\name{set_direc}
\alias{set_direc}
\title{Revise the flow direction of the flow direction data.}
\usage{
set_direc(direc_grid, row, col, direc, arc_code = TRUE, keypad_dir = TRUE)
}
\arguments{
\item{direc_grid}{Flow direction data created by create_flow_direction().}

\item{row}{Row of the gridcell to be revised.}

\item{col}{Column of the gridcell to be revised.}

\item{direc}{New flow direction of the gridcell. For more convinient calibration,
the direction code is sat as the keypad. It means that 1 is southwest,
2 is south, 3 is southeast, and so on. set keypad_dir to FALSE can
revise the direction code directly.}

\item{arc_code}{If use the ArcInfo flow direction code.}

\item{keypad_dir}{If use keypad flow direction code.}
}
\value{
The flow direction data after revise.
}
\description{
Revise the flow direction of the flow direction data created by
create_flow_direction().
}
