% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil_class.R
\name{soil_class}
\alias{soil_class}
\title{soil texture class, based on USDA triangle}
\usage{
soil_class(clay, sand)
}
\arguments{
\item{clay}{Vector, percent of clay}

\item{sand}{Vector, percent of sand}
}
\value{
Character vector of soil texture class, one of
\code{c("Cl", "SiCl", "SaCl", "ClLo", "SiClLo", "SaClLo", "Lo", "SiLo", "SaLo",  "Si", "LoSa", "Sa")}.
}
\description{
soil texture class, based on USDA triangle
}
\examples{
clay = c(05, 60, 15, 05, 25, 05, 25, 45, 65, 75, 13, 47)
sand = c(90, 32, 70, 70, 20, 10, 10, 10, 20, 10, 70, 10)
soil_class(clay, sand)
# c("Sa", "Cl", "SaLo", "SaLo", "SiLo", "Si", "SiLo", "SiCl", "Cl", "Cl", "SaLo", "SiCl")
}
\references{
\enumerate{
\item Julien Moeys (2016). soiltexture: Functions for Soil Texture Plot,
Classification and Transformation. R package version 1.5.1.
https://CRAN.R-project.org/package=soiltexture
}
}
