% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SoilParam.R
\name{SPAW}
\alias{SPAW}
\alias{wilting_point}
\alias{field_capacity}
\alias{saturated_mois}
\alias{wilting_point_salinity}
\title{Soil properties in SPAW}
\usage{
wilting_point(S, C, OM)

field_capacity(S, C, OM)

saturated_mois(S, C, OM)

wilting_point_salinity(S = 0.2, C = 0.2, OM = 2.5, EC = 3)
}
\arguments{
\item{S}{the weight ratio of Sand, (weight ratio, 0-1)}

\item{C}{the weight ratio of Clay, (weight ratio, 0-1)}

\item{OM}{the percentage of Organic Matter, (\%w, 0-100)}

\item{EC}{Electrical conductance of a saturated soil extract (dS/m = mili-mho/cm)}
}
\value{
\%v
}
\description{
Wilting point, field capacity and saturated moisture
\itemize{
\item \verb{wilting point}: 1500 kPa moisture, \%v, (Eq. 1)
\item \verb{field capacity}: 33 kPa moisture, \%v, (Eq. 2)
\item \verb{saturated moisture}: Saturation (0 kPa moisture), \%v, (Eq. 5)
}
}
\note{
The unit of \code{OM} is \verb{\%w}, which is different from \code{S} and \code{C}.
}
\examples{
S = C = 0.20; OM = 2.5 
wilting_point(S, C, OM)  # 13.8\%
field_capacity(S, C, OM) # 32.1\% 
saturated_mois(S, C, OM) # 48.2\%
}
\references{
\enumerate{
\item Saxton, K. E., & Rawls, W. J. (2006). Soil water characteristic estimates
by texture and organic matter for hydrologic solutions. Soil Science
Society of America Journal, 70(5), 1569-1578. \url{doi:10.2136/sssaj2005.0117}
}
}
