% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_sun_angle.R
\name{suncalc}
\alias{suncalc}
\alias{sunrise}
\title{suncalc}
\usage{
suncalc(
  time,
  lon = 114.6053,
  lat = 30.49694,
  ...,
  year = year(Sys.time()),
  verbose = TRUE
)

sunrise(date = Sys.Date(), lon = 80, lat = 30, timeZone = 8)
}
\description{
The default location is ZuoLing.
}
\examples{
suncalc(Sys.Date())

sunrise(Sys.Date())

}
