% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch03_thermo.R
\name{theta_wet2}
\alias{theta_wet2}
\title{theta_wet2}
\usage{
theta_wet2(P0, T0, Td)
}
\description{
theta_wet2
}
\examples{
theta_wet2(850, 20, 18)

}
\references{
\enumerate{
\item https://unidata.github.io/MetPy/latest/api/generated/metpy.calc.equivalent_potential_temperature.html
}
}
