% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit_convert.R
\name{unit_convert}
\alias{unit_convert}
\alias{R2Q}
\alias{Q2R}
\alias{MJ_2W}
\alias{W2_MJ}
\alias{W2mm}
\alias{K2T}
\alias{T2K}
\title{unit convertion}
\usage{
R2Q(R, area = dt * 3.6, dt = 24)

Q2R(Q, area = dt * 3.6, dt = 24)

MJ_2W(x)

W2_MJ(x)

W2mm(x, tmean = 0)

K2T(x)

T2K(x)
}
\arguments{
\item{R}{runoff (mm per dt)}

\item{area}{basin area (km^2), if area = ,}

\item{dt}{time duration (hour)}
}
\description{
\itemize{
\item \code{R2Q}: Q(m^3/s) = R(mm) * area (km^2) * 1000/(3600*24)
\item \code{W2mm}: 1 MJ /m2/day  =0.408 mm /day, 1 Watt /m2 = 0.0864 MJ /m2/day
}
}
