% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil_class.R
\docType{data}
\name{usda_sf}
\alias{usda_sf}
\title{USDA soil class classification}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 12 rows and 2 columns.
}
\usage{
usda_sf
}
\description{
USDA soil class classification
}
\references{
\enumerate{
\item https://en.wikipedia.org/wiki/Soil_texture
}
}
\keyword{datasets}
