% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SoilParam.R
\name{VIC_soilParam}
\alias{VIC_soilParam}
\alias{expt}
\alias{Ksat}
\alias{bubble}
\alias{Wcr_FT}
\alias{Wpwp_FT}
\title{VIC model soil parameters}
\usage{
expt(S, C, OM)

Ksat(S, C, OM)

bubble(S, C, OM)

Wcr_FT(S, C, OM)

Wpwp_FT(S, C, OM)
}
\arguments{
\item{S}{the weight ratio of Sand, (weight ratio, 0-1)}

\item{C}{the weight ratio of Clay, (weight ratio, 0-1)}

\item{OM}{the percentage of Organic Matter, (\%w, 0-100)}
}
\description{
Soil parameters for VIC model (calculated from HWSD database)
\itemize{
\item \code{expt} : 3+2b (Eq. 17)
\item \code{Kstat}: Saturated conductivity (mm/hr) (Eq. 16)
\item \code{bubble}: Bubbling pressure (cm) (Eq. 4)
\item \code{Wcr_FT  }: Field Capacity, Fractional soil moisture content at the
critical point (~70\% of field capacity) (fraction of maximum moisture)
\item \code{Wpwp_FT }: Wilting Point, Fractional soil moisture content at the
wilting point (fraction of maximum moisture)
}
}
\details{
\itemize{
\item Kstat: unit has changed from \code{mm/h} to \code{mm/day}
}
}
\examples{
S = C = 0.20; OM = 2.5
expt(S, C, OM)
Ksat(S, C, OM)           # 12.19
bubble(S, C, OM)
field_capacity(S, C, OM) # 32.1\%
wilting_point(S, C, OM)  # 13.8\%
# Fractional soil moisture
Wcr_FT(S, C, OM)
Wpwp_FT(S, C, OM)
}
\references{
Saxton, K. E., & Rawls, W. J. (2006). Soil water characteristic estimates by
texture and organic matter for hydrologic solutions. Soil Science Society of
America Journal, 70(5), 1569-1578. https://doi.org/10.2136/sssaj2005.0117
}
