% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_XAJ.R
\docType{data}
\name{XAJ.param.range}
\alias{XAJ.param.range}
\title{The lumped XAJ model has 13 parameters, including:}
\format{
An object of class \code{data.frame} with 15 rows and 3 columns.
}
\usage{
XAJ.param.range
}
\description{
\enumerate{
\item KC,   Ratio of potential evap to pan evap
\item IM,   Fraction of impermeable area
\item WUM,  Soil moisture capacity of upper layer
\item WLM,  Soil moisture capacity of lower layer
\item WDM,  Soil moisture capacity of deep layer
\item C,    Coefficient of deep evap
\item B,    Exponent of the soil moisture storage capacity curve
\item SM,   Areal mean free water capacity of the surface soil layer
\item EX,   Exponent of the free water capacity curve
\item KI,   outflow coefficients of the free water storage to interflow
\item KG,   outflow coefficients of the free water storage to groundwater
\item CI,   recession constant of the lower interflow storage
\item CG,   recession constant of groundwater storage.
If use the instantaneous unit hydrograph (IUH) of Nash for routing
of surface runoff, should provided two other parameters:
}
\itemize{
\item N,    number of reservoirs in the instantaneous unit hydrograph
\item NK,   common storage coefficient in the instantaneous unit hydrograph
Else should provided the whole unit hydrograph defined by user,
and set \code{full.UH} become \code{TRUE}.
}
}
\details{
The parameter \code{params} must be a numeric vector looks like:
\code{c(KC, IM, WUM, WLM, WDM, C, B, SM, EX, KI, KG, CI, CG, N, NK)}

when use the instantaneous unit hydrograph of Nash, or looks like:
\code{c(KC, IM, WUM, WLM, WDM, C, B, SM, EX, KI, KG, CI, CG, UH_1, UH_2, ..., UH_n)}
UH_1, UH_2, ..., UH_n means the series of the unit hydrograph.
}
\keyword{datasets}
