% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calib_XAJ.R
\name{XAJ_predict}
\alias{XAJ_predict}
\title{XAJ_predict}
\usage{
XAJ_predict(model, newdata)
}
\arguments{
\item{model}{XAJ model returned by \code{\link[=XAJ_calib]{XAJ_calib()}}}

\item{newdata}{A data.frame, with the column of \code{prcp}, \code{ET0}, \code{date} (optional).}
}
\description{
XAJ_predict
}
