% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_pressure.R
\name{cal_ea}
\alias{cal_ea}
\title{Estimate actual vapor pressure.}
\usage{
cal_ea(tmin, tmax = NULL, RH_max = NULL, RH_min = NULL, RH_mean = NULL)
}
\arguments{
\item{tmin}{Daily minimum air temperature at 2m height \verb{[deg Celsius]}.}

\item{tmax}{Daily maximum air temperature at 2m height \verb{[deg Celsius]}.}

\item{RH_max, RH_mean, RH_min}{Daily max, mean and min relative humidity \verb{[\%]}.}
}
\value{
Actual vapor pressure (i.e. avp or ea) \verb{[kPa]}.
}
\description{
Estimate actual vapor pressure by providing daily maximum and minimum
air temperature, daily maximum, mean and minimum relative humidity.
}
\details{
tmin must be provided. There are 4 options:
\enumerate{
\item \code{RH_max}, \code{RH_min} | \code{tmin}, \code{tmax}: \code{(es(tmax) * RH_min + es(tmin) * RH_max)/200}
\item \code{RH_mean}          | \code{tmin}, \code{tmax}: \code{(es(tmax) + es(tmin)) * RH_mean/200}
\item \code{RH_max}           | \code{tmin}        : \code{es(tmin) * RH_max / 100}
\item \code{tmin}                             : \code{es(tmin)}
}
}
