% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_radiation.R
\name{cal_Ra}
\alias{cal_Ra}
\alias{ext_rad}
\title{Estimate daily daily extraterrestrial radiation.}
\usage{
cal_Ra(lat, dates)

ext_rad(lat, dates)
}
\arguments{
\item{lat}{Latitude \verb{[degree]}.}

\item{dates}{A R Date type of a vector of Date type. If not provided, it will
Regard the ssd series is begin on the first day of a year.}
}
\value{
extraterrestrial radiation \verb{[MJ m-2 day-1]}.
}
\description{
Estimate daily daily extraterrestrial radiation \verb{[MJ m-2 day-1]}.
}
