% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_Rn.R
\name{cal_Rn}
\alias{cal_Rn}
\title{Surface net radiation}
\usage{
cal_Rn(
  lat,
  J,
  Rs = NULL,
  Tmin,
  Tmax,
  ea = NULL,
  RH,
  ssd = NULL,
  cld = NULL,
  albedo = 0.23,
  Z = 0,
  ...
)
}
\arguments{
\item{lat}{float, latitude}

\item{J}{integer, day of year}

\item{Tmin, Tmax}{numeric vector, min and max 2m-air temperature (degC)}

\item{ea}{Vapor pressure}

\item{RH}{Relative humidity (\%). If \code{ea} provided, \code{RH} will be ignored.}

\item{ssd}{numeric vector, sun shine duration (hour)}

\item{cld}{(optional) cloud coverage (0-1). At least one of \code{cld} and \code{ssd}
should be provided. If \code{cld} is not null, \code{ssd} will be ignored.}

\item{albedo}{(optional), \verb{Rsn = (1 - albedo) Rs}}

\item{Z}{(optional) elevation (m), for the calculation of \code{Rso}}
}
\value{
radiation in \verb{[MJ d-1]}
\itemize{
\item \code{Rn}  : Surface net radiation
\item \code{Rln} : Surface outward net longwave radiation (negative means outgoing)
\item \code{Rsn} : Surface downward net shortwave radiation
\item \code{Rs}  : Surface downward shortwave radiation
\item \code{Rso} : Clear-sky surface downward shortwave radiation
\item \code{Ra}  : Extraterrestrial radiation
}
}
\description{
Surface net radiation
}
\note{
\code{Rn} might <= 0. Users need to constrain the min value by \code{pmax(Rn, 0)}.
}
\examples{
cal_Rn(lat = 30, J = 1, RH = 70, Tmin = 20, Tmax = 30, ssd = 10)

}
\author{
Xie YuXuan and Kong Dongdong
}
