% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_Rn.R
\name{cal_Rn}
\alias{cal_Rn}
\title{Surface net radiation}
\usage{
cal_Rn(
  lat,
  J,
  Tmin,
  Tmax,
  ea = NULL,
  RH,
  ssd = NULL,
  cld = NULL,
  Rsi = NULL,
  albedo = 0.23,
  Z = 0,
  ...
)
}
\arguments{
\item{lat}{float, latitude}

\item{J}{integer, day of year}

\item{Tmin, Tmax}{numeric vector, min and max 2m-air temperature (degC)}

\item{ea}{Actual vapor pressure (kPa)}

\item{RH}{Relative humidity (\%). If \code{ea} provided, \code{RH} will be ignored.}

\item{ssd}{numeric vector, sun shine duration (hour)}

\item{cld}{(optional) cloud coverage (0-1). At least one of \code{cld} and \code{ssd}
should be provided. If \code{cld} is not null, \code{ssd} will be ignored.}

\item{Rsi}{(optional) Surface downward shortwave radiation (MJ m-2 d-1).
\itemize{
\item If not provided, \code{Rsi} will be calculated by \code{(as + bs * nN) * Rsi_toa}.
\item If provided, \code{ssd} and \code{cld} will be ignored.
}}

\item{albedo}{(optional), \verb{Rsn = (1 - albedo) Rsi}}

\item{Z}{(optional) elevation (m), for the calculation of \code{Rsi_o}}

\item{...}{ignored}
}
\value{
radiation in \verb{[MJ d-1]}
\itemize{
\item \code{Rn}  : Surface net radiation
\item \code{Rln} : Surface outward net longwave radiation (negative means outgoing)
\item \code{Rsn} : Surface downward net shortwave radiation
\item \code{Rsi}  : Surface downward shortwave radiation (Rsi)
\item \code{Rsi_o} : Clear-sky surface downward shortwave radiation
\item \code{Rsi_toa}  : Extraterrestrial radiation (top of atmosphere, Rsi_toa) #'
}
}
\description{
Surface net radiation
}
\note{
\code{Rn} might <= 0. Users need to constrain the min value by \code{pmax(Rn, 0)}.
}
\examples{
cal_Rn(lat = 30, J = 1, RH = 70, Tmin = 20, Tmax = 30, ssd = 10)
}
\seealso{
\code{\link[=cal_Rsi_toa]{cal_Rsi_toa()}}, \code{\link[=cal_Rsi]{cal_Rsi()}}
}
\author{
Xie YuXuan and Kong Dongdong
}
