% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ET0_helper.R, R/cal_pressure.R
\name{ET0_helper}
\alias{ET0_helper}
\alias{cal_U2}
\alias{cal_ea}
\alias{cal_VPD}
\alias{cal_lambda}
\alias{cal_slope}
\alias{cal_gamma}
\alias{cal_bowen}
\alias{cal_Pa}
\alias{cal_rH}
\alias{cal_rH2}
\alias{cal_TvK}
\alias{cal_rou_a}
\alias{cal_es}
\title{helper functions for potential evapotranspiration (ET0)}
\usage{
cal_U2(Uz, z.wind = 10)

cal_ea(Tmin, Tmax = NULL, RH_mean = NULL)

cal_VPD(Tair, Tdew = NULL)

cal_lambda(Tair)

cal_slope(Tair)

cal_gamma(Tair, Pa = atm)

cal_bowen(Tair, Pa = atm)

cal_Pa(z = NULL)

cal_rH(U2, h = 0.12)

cal_rH2(U2, Tair, Pa = atm)

cal_TvK(Tair, q = NULL, ea = NULL, Pa = atm)

cal_rou_a(Tair, Pa = atm, q = NULL, ea = NULL)

cal_es(Tair)
}
\arguments{
\item{z.wind}{Height where measure the wind speed \verb{[m]}. Default 10m.}

\item{Tmin}{Daily minimum air temperature at 2m height \verb{[deg Celsius]}.}

\item{Tmax}{Daily maximum air temperature at 2m height \verb{[deg Celsius]}.}

\item{Tair}{2m air temperature, (\verb{[deg Celsius]})}

\item{Tdew}{dew temperature, (\verb{[deg Celsius]})}

\item{Pa}{land surface Air pressure \verb{[kPa]}.}

\item{z}{Elevation above sea level \verb{[m]}. Must provided if pres
are not provided.}

\item{RH_max, RH_mean, RH_min}{Daily max, mean and min relative humidity \verb{[\%]}.}
}
\description{
\itemize{
\item \code{lambda}: latent heat of vaporization, about \verb{[2.5 MJ kg-1]}.
\item \code{slope}: The slope of the saturation vapour pressure curve at certain air
temperature \code{Tair}, \verb{[kPa degC-1]}.
\item \code{gamma}: psychrometric constant (\verb{[kPa degC-1]}), \code{Cp*Pa/(epsilon*lambda)}.
\item \code{U2}: 10m wind speed (m/s). According to wind profile relationship, convert U_z to U_2.
\item \code{es}: saturation vapor pressure (kPa)
\item \code{ea}: actual vapor pressure (kPa)
\enumerate{
\item \code{RH_mean} | \code{Tmin}, \code{Tmax}: \code{(es(Tmax) + es(Tmin)) * RH_mean/200}
\item \code{Tmin}                    : \code{es(Tmin)}
}
\item \code{cal_TvK}: vitual temperature (K)
\enumerate{
\item Tair + q
\item Tair + ea/Pa
\item \code{1.01 * (Tair + 273)} , FAO56 Eq. 3-7
}
}
}
\examples{
cal_VPD(10, 5)

par(mfrow = c(2, 2), mar = c(3, 1.8, 2, 1), mgp = c(2, 0.6, 0))
Tair = -10:50
plot(cal_es(-10:50), main = "es (kPa)", xlab = "Tair")
plot(cal_bowen(-10:50), main = "Bowen ratio", xlab = "Tair")
plot(cal_slope(-10:50), main = "slope (kPa/degC)", xlab = "Tair")
plot(cal_gamma(-10:50), main = "gamma (kPa/degC)", xlab = "Tair")
}
\references{
\enumerate{
\item Allen, R. G., & Luis S. Pereira. (1998). Crop
evapotranspiration-Guidelines for computing crop water requirements-FAO
Irrigation and drainage paper 56. European Journal of Agronomy, 34(3),
144–152. doi:10.1016/j.eja.2010.12.001
}
}
