% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ET0_models.R
\name{ET0_models}
\alias{ET0_models}
\alias{ET0_eq}
\alias{ET0_Penman48}
\alias{ET0_Monteith65}
\alias{ET0_PT72}
\alias{ET0_FAO98}
\title{Potential Evapotranspiration models}
\usage{
ET0_eq(Rn, Tair, Pa = atm, ...)

ET0_Penman48(Rn, Tair, Pa = atm, D, wind, z.wind = 10)

ET0_Monteith65(Rn, Tair, Pa = atm, D, wind, z.wind = 10, rs = 70, ...)

ET0_PT72(Rn, Tair, Pa = atm, alpha = 1.26, ...)

ET0_FAO98(Rn, Tair, Pa = atm, D, wind, z.wind = 10, tall_crop = FALSE)
}
\arguments{
\item{Rn}{land surface net radiation, W m-2}

\item{Tair}{2m air temperature (degC)}

\item{Pa}{surface air pressure (kPa)}

\item{D}{vapor pressure deficit (kPa)}

\item{wind}{wind speed at the height of \code{z.wind}}

\item{z.wind}{Height where measure the wind speed \verb{[m]}. Default 10m.}
}
\description{
\itemize{
\item \code{ET0_eq}   : Equilibrium evaporation, \code{slope / (slope + gamma) * Rn}
\item \code{ET0_Penman48} : Penman 1948 equation simplified by Shuttleworth 1993
\item \code{ET0_Monteith65}: Penman-Monteith 1965
\item \code{ET0_PT72} : Priestley Taylor 1972, \code{ET0_PT72 = ET0_eq * 1.26}
\item \code{ET0_FAO98} : Penman-Monteith reference crop evapotranspiration, FAO56
}
}
\examples{
et0_pt = ET0_PT72(250, 25, D = 1)
et0_penman48 = ET0_Penman48(250, 25, D = 1, wind = 2)
et0_fao98 = ET0_FAO98(250, 25, D = 1, wind = 2)
}
\references{
\enumerate{
\item Penman equation. Wikipedia 2021.
\url{https://en.wikipedia.org/w/index.php?title=Penman_equation}
\item Shuttleworth, W. J. (1993) Evaporation. In: Handbook of Hydrology (ed. by
D. Maidment). McGraw-Hill, New York.
\item Allen, R. G., & Luis S. Pereira. (1998). Crop
evapotranspiration-Guidelines for computing crop water requirements-FAO
Irrigation and drainage paper 56. European Journal of Agronomy, 34(3),
144-152. \url{doi:10.1016/j.eja.2010.12.001}
\item Allen, R. G., Pereira, L. S., Raes, D., & Smith, M. (1998). FAO Irrigation
and drainage paper No. 56. Rome: Food and Agriculture Organization of the
United Nations, 56(97), e156.
\item Chapter 2 - FAO Penman-Monteith equation,
\url{https://www.fao.org/3/x0490E/x0490e06.htm}
}
}
