% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flood_divide.R
\name{detect_groups}
\alias{detect_groups}
\alias{detect_flood_events}
\alias{flood_divide}
\alias{merge_flood}
\title{detect_flood_events}
\usage{
detect_groups(df, inds, gap_max = 5, extend = ddays(5))

detect_flood_events(
  date,
  Q,
  Q_min = 2,
  Q_peak = 10,
  gap_max = 5,
  extend = ddays(5)
)

flood_divide(df, ...)

merge_flood(df, info_flood, format = "\%Y.\%m")
}
\arguments{
\item{df}{A data.table with date and Q columns}

\item{gap_max}{Default \code{5} days.
\itemize{
\item \code{gaps} (in days) <= \code{gap_max} is regarded as the same event.
\item \code{gaps > gap_max}, it will be regarded as two events.
}}

\item{extend}{Default \code{ddays(5)}. Extend \code{nday} in the left and right of a event}

\item{Q_min}{minimum discharge to detect flood events}

\item{Q_peak}{peak discharge to detect flood events}

\item{...}{parameters passed to \code{\link[=detect_flood_events]{detect_flood_events()}}}
}
\description{
detect_flood_events

flood_divide
}
