% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch03_thermo.R
\name{LCL}
\alias{LCL}
\title{Lifted Condensation Level (LCL)}
\usage{
LCL(P0, T0, Td)
}
\arguments{
\item{P0}{pressure at surface (hPa)}

\item{T0}{temperature at surface (K)}

\item{Td}{dew point temperature (K)}
}
\description{
Lifted Condensation Level (LCL)
}
