% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ET0_penman65.R
\name{PET_hg}
\alias{PET_hg}
\title{Estimate ET0 by Hargreaves equation.}
\usage{
PET_hg(Tmax, Tmin, Tavg = NULL, Ra = NULL, lat = NULL, dates = NULL)
}
\arguments{
\item{Tmax}{Daily maximum air temperature at 2m height \verb{[deg Celsius]}.}

\item{Tmin}{Daily minimum air temperature at 2m height \verb{[deg Celsius]}.}

\item{Tavg}{Daily mean air temperature at 2m height \verb{[deg Celsius]}. If not
provided it would estimated by averaging the Tmax and tmin.}

\item{Ra}{Clear sky incoming shortwave radiation, i. e. extraterrestrial
radiation multiply by clear sky transmissivity (i. e. a + b, a and b are
coefficients of Angstrom formula. Normally 0.75) \verb{[MJ m-2 day-1]}. If not
provided, must provide lat and dates.}

\item{lat}{Latitude \verb{[degree]}.}

\item{dates}{A R Date type of a vector of Date type. If not provided, it will Regard
the ssd series is begin on the first day of a year.}
}
\value{
Reference evapotranspiration ET0 from a hypothetical grass reference
surface \verb{[mm day-1]}.
}
\description{
Estimate reference evapotranspiration (ET0) from a hypothetical
short grass reference surface using the the Hargreaves equation.
}
