% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calib_XAJ.R
\name{plot_calib}
\alias{plot_calib}
\title{plot_calib}
\usage{
plot_calib(data, ...)
}
\arguments{
\item{data}{with the columns of \code{date}, \code{prcp}, \code{Qsim},  \code{Qobs}}

\item{...}{others to plot precipitation}
}
\description{
plot_calib
}
