% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_runoff.R
\name{plot_runoff}
\alias{plot_runoff}
\title{plot_runoff}
\usage{
plot_runoff(
  df_q,
  df_prcp = NULL,
  xlim = NULL,
  ylim2 = c(50, 0),
  legend.position = c(1, 1),
  legend.justification = c(1, 1),
  ...
)
}
\arguments{
\item{df_q}{A data.frame with the columns of \code{date}, \code{Qsim}, \code{Qobs}.
At least one of \code{Qsim} and \code{Qobs} should be provided.}

\item{df_prcp}{A data.frame with the columns of \code{date}, \code{prcp}}

\item{xlim}{limit of x axis}

\item{ylim2}{limit of second y axis (for precipitation)}

\item{...}{other parameters passed to \code{\link[=plot]{plot()}} for precipitation plot}
}
\description{
plot_runoff
}
