% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit_convert.R
\name{unit_convert}
\alias{unit_convert}
\alias{R2Q}
\alias{Q2R}
\alias{MJ_2W}
\alias{MJ_2mm}
\alias{W2_MJ}
\alias{W2mm}
\alias{K2T}
\alias{T2K}
\alias{deg2rad}
\alias{rad2deg}
\title{unit convertion}
\usage{
R2Q(R, area = dt * 3.6, dt = 24)

Q2R(Q, area = dt * 3.6, dt = 24)

MJ_2W(x)

MJ_2mm(x)

W2_MJ(x)

W2mm(x, tmean = 0)

K2T(x)

T2K(x)

deg2rad(deg)

rad2deg(rad)
}
\arguments{
\item{R}{runoff (mm per dt)}

\item{area}{basin area (km^2), if area = ,}

\item{dt}{time duration (hour)}

\item{Q}{runoff flow (m^3/s)}

\item{x}{scalar or numeric vector}

\item{tmean}{daily mean temperature}

\item{deg}{angle in degree}

\item{rad}{angle in radian}
}
\description{
\itemize{
\item \code{R2Q}: Q(m^3/s) = R(mm) * area (km^2) * 1000/(3600*24)
\item \code{W2mm}: 1 MJ /m2/day  =0.408 mm /day, 1 Watt /m2 = 0.0864 MJ /m2/day
}
}
\references{
http://www.fao.org/3/X0490E/x0490e0i.htm
}
