% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_pressure.R
\name{vapour_press}
\alias{vapour_press}
\alias{q2ea}
\alias{ea2VPD}
\alias{q2RH}
\alias{RH2q}
\alias{w2q}
\alias{q2w}
\title{Helper functions for vapour pressure}
\usage{
q2ea(q, Pa)

ea2VPD(ea, RH)

vapour_press(q, Pa)

q2RH(q, Pa, Tair)

RH2q(RH, Pa, Tair)

w2q(w, Pa)

q2w(q, Pa)
}
\arguments{
\item{q}{specific humidity in kg/kg or g/g}

\item{Pa}{surface air pressure}

\item{RH}{relative humidity, in \%}

\item{Tair}{air temperature, in degC}

\item{w}{mix ratio, m_w / m_d}
}
\value{
e in the same unit as Pa
}
\description{
Helper functions for vapour pressure
}
\examples{
RH = 90
Pa = atm
Tair = 30
q = RH2q(RH, Pa, Tair)
RH2 = q2RH(q, Pa, Tair)
e = q2ea(q, Pa)
es = cal_es(Tair)
}
\references{
https://www.eol.ucar.edu/projects/ceop/dm/documents/refdata_report/eqns.html, Eq-17

https://earthscience.stackexchange.com/questions/2360/how-do-i-convert-specific-humidity-to-relative-humidity
}
