% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_pressure.R
\name{vapour_press}
\alias{vapour_press}
\alias{q2ea}
\alias{q2RH}
\alias{RH2q}
\alias{w2q}
\alias{q2w}
\title{Actual vapour pressure (kPa)}
\usage{
vapour_press(q, p)

q2ea(q, p)

q2RH(q, p, Tair)

RH2q(RH, p, Tair)

w2q(w, p)

q2w(q, p)
}
\arguments{
\item{q}{specific humidity in kg/kg or g/g}

\item{p}{surface air pressure}

\item{Tair}{air temperature, in degC}

\item{RH}{relative humidity, in \%}

\item{w}{mix ratio, m_w / m_d}
}
\value{
e in the same unit as p
}
\description{
Actual vapour pressure (kPa)
}
\examples{
RH = 90
p = atm
Tair = 30
q = RH2q(RH, p, Tair)
RH2 = q2RH(q, p, Tair)
e = q2ea(q, p)
es = cal_es(Tair)
}
\references{
https://www.eol.ucar.edu/projects/ceop/dm/documents/refdata_report/eqns.html, Eq-17

https://earthscience.stackexchange.com/questions/2360/how-do-i-convert-specific-humidity-to-relative-humidity
}
