% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_pressure.R
\name{vapour_press}
\alias{vapour_press}
\alias{q2ea}
\alias{q2RH}
\alias{RH2q}
\alias{cal_es}
\title{Actual vapour pressure (kPa)}
\usage{
vapour_press(q, p)

q2ea(q, p)

q2RH(q, p, t)

RH2q(RH, t)

cal_es(t)
}
\arguments{
\item{q}{specific humidity in kg/kg}

\item{p}{surface air pressure}

\item{t}{Air temperature \verb{[Celsius]}.}
}
\value{
e in the same unit as p

saturation vapor pressure \verb{[kPa]}.
}
\description{
Estimate saturation vapor pressure from air temperature.
}
\references{
https://www.eol.ucar.edu/projects/ceop/dm/documents/refdata_report/eqns.html, Eq-17
}
