% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_pressure.R
\name{vapour_press}
\alias{vapour_press}
\alias{w2q}
\alias{q2w}
\alias{q2ea}
\alias{w2ea}
\alias{ea2VPD}
\alias{q2RH}
\alias{q_from_RH}
\alias{RH2q}
\alias{Tdew2q}
\alias{Tdew2w}
\alias{Tdew2RH}
\alias{Tdew_from_q}
\alias{Tdew_from_w}
\title{Helper functions for vapour pressure}
\usage{
w2q(w)

q2w(q, Pa)

q2ea(q, Pa)

w2ea(w, Pa)

ea2VPD(ea, RH)

vapour_press(q, Pa)

q2RH(q, Tair, Pa)

q_from_RH(RH, Tair, Pa)

RH2q(RH, Tair, Pa)

Tdew2q(Tdew, Pa)

Tdew2w(Tdew, Pa)

Tdew2RH(Tdew, Tair)

Tdew_from_q(q, Pa)

Tdew_from_w(w, Pa)
}
\arguments{
\item{w}{mix ratio, m_w / m_d}

\item{q}{specific humidity in kg/kg or g/g}

\item{Pa}{surface air pressure}

\item{ea}{actual vapor pressure (kPa)}

\item{RH}{relative humidity, in \%}

\item{Tair}{air temperature, in degC}

\item{Tdew}{dew temperature (in degC)}
}
\value{
e in the same unit as Pa
}
\description{
Helper functions for vapour pressure
}
\examples{
RH = 90
Pa = atm
Tair = 30
q = RH2q(RH, Pa, Tair)
RH2 = q2RH(q, Pa, Tair)
e = q2ea(q, Pa)
es = cal_es(Tair)
}
\references{
https://www.eol.ucar.edu/projects/ceop/dm/documents/refdata_report/eqns.html, Eq-17

https://earthscience.stackexchange.com/questions/2360/how-do-i-convert-specific-humidity-to-relative-humidity
}
